move AMDSScalerDetectorManager class into AMDSScalerDetector.h
