+++ source/Detector/Scaler/AMDSScalerDetector.cpp
@@ -13,6 +13,30 @@
 
 #include "util/AMErrorMonitor.h"
 
+/// ==================== implementation of AMDSScalerDetectorManager ============================
+AMDSScalerDetectorManager::AMDSScalerDetectorManager(AMDSScalerConfigurationMap *scalerConfiguration, QObject *parent)
+	: QObject(parent)
+{
+	detectorManagerThread_ = new QThread();
+
+	scalerDetector_ = new AMDSScalerDetector(scalerConfiguration);
+	scalerDetector_->moveToThread(detectorManagerThread_);
+
+	connect(detectorManagerThread_, SIGNAL(finished()), scalerDetector_, SLOT(deleteLater()));
+
+	detectorManagerThread_->start();
+}
+
+AMDSScalerDetectorManager::~AMDSScalerDetectorManager()
+{
+	if (detectorManagerThread_->isRunning())
+		detectorManagerThread_->quit();
+
+	detectorManagerThread_->deleteLater();
+	detectorManagerThread_ = 0;
+}
+
+/// ==================== implementation of AMDSScalerDetector ============================
 AMDSScalerDetector::AMDSScalerDetector(AMDSScalerConfigurationMap *scalerConfiguration, QObject *parent)
     : QObject(parent)
 {
